+++ azure-mgmt-datalake-analytics/azure/mgmt/datalake/analytics/account/models/compute_policy_paged.py
@@ -9,5 +9,19 @@
 # regenerated.
 # --------------------------------------------------------------------------
 
-VERSION = "0.1.5"
+from msrest.paging import Paged
 
+
+class ComputePolicyPaged(Paged):
+    """
+    A paging container for iterating over a list of ComputePolicy object
+    """
+
+    _attribute_map = {
+        'next_link': {'key': 'nextLink', 'type': 'str'},
+        'current_page': {'key': 'value', 'type': '[ComputePolicy]'}
+    }
+
+    def __init__(self, *args, **kwargs):
+
+        super(ComputePolicyPaged, self).__init__(*args, **kwargs)